requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
ebooklib>=0.18
markdown>=3.4.0
weasyprint>=60.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
